Change custom matchers to be extensible/readable
The old ":Should {actual} be {foo}" syntax is very limited.  It is not
possible to pass extra arguments to custom matchers, and it is less
readable because :Should has been renamed as :Expected.

So that the custom matcher syntax is changed as follows:
":Expect {actual} to{foo} [{arg}, ...]".
It is extensible and more readable.