allow a comment at the end of line
Vim script seems to allow a comment at the end of line.
But `describe`, `it` and other vspec grammer don't allow it.
I modified a regexp for parsing and make it allow a comment at the end
of line.

I implemented this feature because I want to fold a block with `" {{{`
like below.

```vim
describe 'hoge()' " {{{

    it 'returns "hoge"' " {{{
    end
    " }}}
end
" }}}
```