Fix not to load test script twice
If bin/vspec is invoked with no dependency, a test script is read twice:

(1) Vim loads the a script into buffer #1.
(2) Bootstrap script compiles the test script.

This is not desired.  If a test script is a process substitution, the
whole content is consumed by (1).  So that bootstrap script cannot read
the content to be compiled.

"--cmd 'source FILE'" should be used to run bootstrap script before
loading any file by Vim.